Move img_utils into contrib folder
