+++ autogen/agentchat/contrib/img_utils.py
